Fix: Add half-star rating system with drag interaction
Implemented half-star ratings (0.5 increments) to provide more granular
book rating feedback. Users can now drag across stars for precise
ratings from 0.5 to 5.0 stars.

Changes:
- Updated data models (BookModel, GoogleBooksAPI) from Int to Double
- Added backward-compatible migration for existing integer ratings
- Implemented smooth drag gesture across stars for rating input
- Added half-star icon display (star.leadinghalf.filled) throughout app
- Updated all import/export services to handle Double ratings
- Polished drag interaction with haptic feedback on rating changes

+++ Epilogue/Epilogue/Models/BookModel.swift
@@ -40,7 +40,7 @@ final class BookModel {
     var isInLibrary: Bool = false  // Default for CloudKit
     var readingStatus: String = ReadingStatus.wantToRead.rawValue // Store as string for SwiftData
     var currentPage: Int = 0  // Default for CloudKit
-    var userRating: Int?
+    var userRating: Double?  // Supports half-star ratings (0.5 increments: 1.0, 1.5, 2.0, etc.)
     var userNotes: String?
     var dateAdded: Date = Date()  // Default for CloudKit
